requirements.txt
# StealthMark Dependencies
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
imageio-ffmpeg>=0.4.9
imageio>=2.31.0
cryptography>=41.0.0
# Optional: HEIC support (requires system libheif)
# pillow-heif>=0.14.0
